Organise data frame and vectors code
