+++ packages/grafana-data/src/dataframe/index.ts
@@ -1,0 +1,4 @@
+export * from './DataFrameView';
+export * from './FieldCache';
+export * from './CircularDataFrame';
+export * from './MutableDataFrame';
