+++ packages/grafana-data/src/index.ts
@@ -1,2 +1,4 @@
 export * from './utils/index';
 export * from './types/index';
+export * from './vector';
+export * from './dataframe';
