+++ packages/grafana-data/src/types/dataFrame.ts
@@ -4,6 +4,7 @@ import { QueryResultBase, Labels, NullValueMode } from './data';
 import { FieldCalcs } from '../utils/index';
 import { DisplayProcessor } from './displayValue';
 import { DataLink } from './dataLink';
+import { Vector } from './vector';
 
 export enum FieldType {
   time = 'time', // or date
@@ -42,25 +43,6 @@ export interface FieldConfig {
 
   // Alternative to empty string
   noValue?: string;
-}
-
-export interface Vector<T = any> {
-  length: number;
-
-  /**
-   * Access the value by index (Like an array)
-   */
-  get(index: number): T;
-
-  /**
-   * Get the resutls as an array.
-   */
-  toArray(): T[];
-
-  /**
-   * Return the values as a simple array for json serialization
-   */
-  toJSON(): any; // same results as toArray()
 }
 
 export interface Field<T = any, V = Vector<T>> {
