+++ packages/grafana-data/src/utils/index.ts
@@ -12,9 +12,6 @@ export * from './object';
 export * from './moment_wrapper';
 export * from './thresholds';
 export * from './text';
-export * from './dataFrameHelper';
-export * from './dataFrameView';
-export * from './vector';
 
 export { getMappedValue } from './valueMappings';
