+++ packages/grafana-data/src/vector/ConstantVector.ts
@@ -1,0 +1,22 @@
+import { Vector } from '../types/vector';
+
+export class ConstantVector<T = any> implements Vector<T> {
+  constructor(private value: T, private len: number) {}
+
+  get length() {
+    return this.len;
+  }
+
+  get(index: number): T {
+    return this.value;
+  }
+
+  toArray(): T[] {
+    const arr = new Array<T>(this.length);
+    return arr.fill(this.value);
+  }
+
+  toJSON(): T[] {
+    return this.toArray();
+  }
+}
