+++ packages/grafana-data/src/vector/vectorToArray.ts
@@ -1,0 +1,9 @@
+import { Vector } from '../types/vector';
+
+export function vectorToArray<T>(v: Vector<T>): T[] {
+  const arr: T[] = [];
+  for (let i = 0; i < v.length; i++) {
+    arr[i] = v.get(i);
+  }
+  return arr;
+}
